Cleanup and refactor out some common classes
